requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
idna==3.3
PySimpleGUI==4.60.3
python-gedcom==1.0.0
PyYAML==6.0
requests==2.28.1
urllib3==1.26.11
